+++ src/main/java/org/folio/circulation/infrastructure/storage/users/AddressTypeRepository.java
@@ -1,4 +1,4 @@
-package org.folio.circulation.infrastructure.storage;
+package org.folio.circulation.infrastructure.storage.users;
 
 import static org.folio.circulation.support.Result.ofAsync;
 import static org.folio.circulation.support.Result.succeeded;
