+++ src/main/java/org/folio/circulation/resources/RequestCollectionResource.java
@@ -24,7 +24,7 @@ import org.folio.circulation.domain.UpdateRequestQueue;
 import org.folio.circulation.domain.UpdateRequestService;
 import org.folio.circulation.domain.UpdateUponRequest;
 import org.folio.circulation.domain.UserManualBlock;
-import org.folio.circulation.infrastructure.storage.UserRepository;
+import org.folio.circulation.infrastructure.storage.users.UserRepository;
 import org.folio.circulation.domain.notice.schedule.RequestScheduledNoticeService;
 import org.folio.circulation.infrastructure.storage.LoanPolicyRepository;
 import org.folio.circulation.infrastructure.storage.RequestPolicyRepository;
