+++ src/main/java/org/folio/circulation/resources/renewal/RenewByBarcodeResource.java
@@ -6,7 +6,7 @@ import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.infrastructure.storage.LoanRepository;
-import org.folio.circulation.infrastructure.storage.UserRepository;
+import org.folio.circulation.infrastructure.storage.users.UserRepository;
 import org.folio.circulation.storage.SingleOpenLoanByUserAndItemBarcodeFinder;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
